+++ src/main/java/nova/backend/domain/user/controller/AuthApi.java
@@ -1,4 +1,4 @@
-package nova.backend.user.controller;
+package nova.backend.domain.user.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -9,9 +9,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import nova.backend.global.common.SuccessResponse;
-import nova.backend.user.dto.request.UserLoginRequestDTO;
-import nova.backend.user.dto.request.UserTokenRequestDTO;
-import nova.backend.user.dto.response.UserTokenResponseDTO;
+import nova.backend.domain.user.dto.request.UserLoginRequestDTO;
+import nova.backend.domain.user.dto.request.UserTokenRequestDTO;
+import nova.backend.domain.user.dto.response.UserTokenResponseDTO;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
