+++ src/main/java/nova/backend/domain/user/dto/request/UserLoginRequestDTO.java
@@ -1,8 +1,8 @@
-package nova.backend.user.dto.request;
+package nova.backend.domain.user.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import nova.backend.user.entity.SocialType;
+import nova.backend.domain.user.entity.SocialType;
 
 public record UserLoginRequestDTO(
         @NotBlank String code,
