+++ src/main/java/nova/backend/domain/user/dto/request/UserUpdateRequestDTO.java
@@ -1,6 +1,6 @@
-package nova.backend.user.dto.request;
+package nova.backend.domain.user.dto.request;
 
-import nova.backend.user.entity.User;
+import nova.backend.domain.user.entity.User;
 
 public record UserUpdateRequestDTO(
         String phoneNumber,
