+++ src/main/java/nova/backend/domain/user/entity/Role.java
@@ -1,4 +1,4 @@
-package nova.backend.user.entity;
+package nova.backend.domain.user.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
