+++ src/main/java/nova/backend/domain/user/entity/SocialType.java
@@ -1,4 +1,4 @@
-package nova.backend.user.entity;
+package nova.backend.domain.user.entity;
 
 public enum SocialType {
     KAKAO, GOOGLE
