+++ src/main/java/nova/backend/domain/user/repository/UserRepository.java
@@ -1,7 +1,7 @@
-package nova.backend.user.repository;
+package nova.backend.domain.user.repository;
 
-import nova.backend.user.entity.SocialType;
-import nova.backend.user.entity.User;
+import nova.backend.domain.user.entity.SocialType;
+import nova.backend.domain.user.entity.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
