+++ src/main/java/nova/backend/domain/user/service/OauthTokenProvider.java
@@ -1,9 +1,9 @@
-package nova.backend.user.service;
+package nova.backend.domain.user.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import nova.backend.global.error.exception.BusinessException;
-import nova.backend.user.entity.SocialType;
+import nova.backend.domain.user.entity.SocialType;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
